example/requirements.txt
Django==2.0.5
django-suit==0.2.28
pypugjs==5.8.1
django-tables2==1.21.2
django-formtools==2.1